Fix detection of SelfSignedTLSClientAuthentication
Self signed TLS is now chosen with the subject DN matches the root DN.
When the PKI method is used, the request will have a certificate, but
its root DN will be null (or at least not equal to the subject DN). The
tests are updated to correctly setup the request and assert the updated
behavior.